split source into modules
