+++ app/Error.hs
@@ -1,0 +1,8 @@
+
+
+module Error where
+
+class Error a where
+    noMsg :: a
+    strMsg :: String -> a
+
